requirements.txt
requests
BeautifulSoup4
python-dotenv
pytz
pymongo
dnspython
